fix(cli): extract workspace-migrate from frozen operator_cli
Keep core/src/operator_cli.rs at 11655 lines so the v0.36.26 desktop-split budget and wiring gate pass without raising the freeze.
