+++ core/src/main.rs
@@ -33,6 +33,7 @@ mod dogfood;
 mod machine_contract;
 mod context_pack;
 mod operator_cli;
+mod workspace_migrate;
 mod project_settings_render;
 mod client;
 mod app;
@@ -936,7 +937,9 @@ fn run_main() -> io::Result<()> {
         "desktop-summary" => return operator_cli::run_desktop_summary_command(&cmd_args[1..]),
         "meta-plan" => return operator_cli::run_meta_plan_command(&cmd_args[1..]),
         "workspace-plan" => return operator_cli::run_workspace_plan_command(&cmd_args[1..]),
-        "workspace-migrate" => return operator_cli::run_workspace_migrate_command(&cmd_args[1..]),
+        "workspace-migrate" => {
+            return workspace_migrate::run_workspace_migrate_command(&cmd_args[1..])
+        }
         "provider-capabilities" => return operator_cli::run_provider_capabilities_command(&cmd_args[1..]),
         "operator-jobs" => return operator_cli::run_operator_jobs_command(&cmd_args[1..]),
         "skills" => return operator_cli::run_skills_command(&cmd_args[1..]),
